move gprc service class to client/service dir
