+++ src/client/CMakeLists.txt
@@ -1,4 +1,4 @@
-file(GLOB_RECURSE CLIENT_SOURCE CONFIGURE_DEPENDS ./*.cc)
+file(GLOB_RECURSE CLIENT_SOURCE CONFIGURE_DEPENDS ./*.cc ./service/*.cc)
 file(GLOB_RECURSE CORE_SOURCE CONFIGURE_DEPENDS ../core/*.cc)
 
 add_library(
